+++ src/features/auth/components/ui/check/buttons/GoogleLoginButton.tsx
@@ -2,7 +2,7 @@ import { useDispatch } from 'react-redux'
 import { openCenteredPopup } from 'src/shared/utils/popup'
 import { setGooglePopup } from 'src/features/auth/utils/popupManager'
 import ButtonGoogle from 'src/assets/icons/button/bbangle-google-button.svg?react'
-import { GOOGLE } from 'src/features/auth/locales/socialLogin'
+import { GOOGLE } from 'src/features/auth/locales/socialProvider'
 import { authAction } from 'src/features/auth/store/authReducer'
 
 const GoogleLoginButton = () => {
