+++ src/features/auth/components/ui/check/buttons/KakaoLoginButton.tsx
@@ -2,7 +2,7 @@ import { useDispatch } from 'react-redux'
 import { openCenteredPopup } from 'src/shared/utils/popup'
 import { setKakaoPopup } from 'src/features/auth/utils/popupManager'
 import ButtonKakao from 'src/assets/icons/button/bbangle-kakao-button.svg?react'
-import { KAKAO } from 'src/features/auth/locales/socialLogin'
+import { KAKAO } from 'src/features/auth/locales/socialProvider'
 import { authAction } from 'src/features/auth/store/authReducer'
 
 const KakaoLoginButton = () => {
