+++ src/features/auth/locales/socialProvider.ts
